requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14
requests>=2.31
python-dotenv>=1.0.1
gunicorn>=21.2.0  # For deployment
